refactor(usage): move the breakdown in with the page it belongs to
It is a section of the new dashboard, so it sits in pages/usage/components
alongside UsageMeter and UsageOverTime rather than in the folder the old page
is being deleted from.
