+++ frontend/web/components/pages/usage/components/UsageBreakdown/components/ScopeTotal.tsx
